+++ packages/app/src/runner/utils.ts
@@ -2,7 +2,7 @@ export const RUNNER_ID = 'unified-runner'
 
 export const REPORTER_ID = 'unified-reporter'
 
-export const MESSAGE_ID = 'unified-messages'
+export const SNAPSHOT_CONTROLS_ID = 'unified-messages'
 
 function getElementById (id: string) {
   const el = document.querySelector<HTMLElement>(`#${id}`)
